Added Xcode project for PPMd code, improved Fuzzer, and fixed bugs in PPMd code, which now doesn't seem to crash.

+++ PPMd/SubAllocatorBrimstone.h
@@ -7,7 +7,7 @@ typedef struct PPMdSubAllocatorBrimstone
 {
 	PPMdSubAllocator core;
 
-	long SubAllocatorSize;
+	uint32_t SubAllocatorSize;
 	uint8_t Index2Units[38],Units2Index[128];
 	uint8_t *LowUnit,*HighUnit;
 	struct PPMAllocatorNodeBrimstone { struct PPMAllocatorNodeBrimstone *next; } FreeList[38];
